fix: styudyroom.js 내용 api폴더로 옮김
