+++ server/api-gateway/routes/api/studyroom/index.js
@@ -1,0 +1,24 @@
+const express = require("express");
+const router = express.Router();
+
+const { makePacket } = require("../../../../lib/tcp/util");
+
+module.exports = function(apigateway) {
+  router.get("/availableRooms", (req, res, next) => {
+    const studyGroupData = req.body;
+
+    req.packet = makePacket(
+      "GET",
+      "apigateway",
+      "availableRooms",
+      "filterStudyGroup",
+      { ...studyGroupData },
+      {},
+      req.resKey,
+      apigateway.context
+    );
+
+    next();
+  });
+  return router;
+};
